refactor: レイヤー変換の共通部分を convert.py に一本化する
convert_vector / convert_raster は種別に依存しない処理をそれぞれ抱えていて、
同じロジックの実装が2つある状態だった:

- クリック入口（42行 vs 43行、実質の差は refresh_kumoy_browser の1行だけ）
- プロジェクトindex解決（raster は関数、vector はインライン）
- レイヤーツリーの差し替え（同上）
- _copy_layer_style（差は _repair_nan_classification の1行だけ）

ui/layers/convert.py に入口・オーケストレーション・上記の共通処理を集約し、
種別ごとに本当に違う「アップロードしてKumoyレイヤーを作る」部分だけを
_upload_vector / _upload_raster に残す。種別分岐は convert_layer_to_kumoy の
isinstance 1箇所だけになる。

実行モデルの違い（ベクター＝メインスレッド同期実行、ラスター＝ワーカースレッド）
はそのまま各モジュールに閉じている。
